+++ javascripts/controllers/manageCourse.js
