+++ router/logout.js
