+++ router/manage.js
