+++ templates/manageCourse.mustache
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+﻿<!DOCTYPE html>
 
 <html lang="en">
 
@@ -10,7 +10,7 @@
 
     <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/1.9.0/jquery.min.js"></script>
     <script type="text/javascript" src="https://rawgit.com/enyo/dropzone/master/dist/dropzone.js"></script>
-    <script type="text/javascript" src="/routers/routes/manage.js"></script>
+    <script type="text/javascript" src="/routers/manage.js"></script>
     <script type="text/javascript" src="/javascripts/controllers/manageCourse.js"></script>
     <link rel="stylesheet"  type="text/css" href="/stylesheets/global.css">
 <link rel="stylesheet" type="text/css" href="/stylesheets/manageCourse.css" />
